+++ src/renderer/renderer.js
@@ -1,8 +1,3 @@
-/**
- * @license
- * Copyright 2023 GappleCider
- * MIT
- */
 class Renderer {
   constructor() {
     this.objs = [];
@@ -129,157 +124,4 @@ class Renderer {
   }
 }
 
-class Vec2 {
-  constructor(x = 0, y = 0) {
-    this.x = x;
-    this.y = y;
-  }
-  get xPos() {
-    return this.x
-  }
-  get yPos() {
-    return this.y
-  }
-  set xPos(x) {
-    this.x = x;
-
-    return this
-  }
-  set yPos(y) {
-    this.y = y;
-
-    return this
-  }
-  set(x, y) {
-    this.x = x;
-    this.y = y;
-
-    return this
-  }
-  setPostion(v) {
-    this.x = v;
-    this.y = v;
-
-    return this
-  }
-  clone() {
-    return this.constructor(this.x, this.y);
-  }
-  scale(s) {
-    this.x *= s;
-    this.y *= s;
-
-    return this
-  }
-  getPosition() {
-    const pos = {
-        x: this.x,
-        y: this.y
-    };
-
-    return pos
-  }
-}
-
-class Polygon {
-  constructor(points = []) {
-    this.props = {
-      visible: true,
-    };
-
-    this.polygon = points;
-    this.tags = {};
-
-    this.class = "polygon";
-
-    this.points.forEach((point) => {
-      point.push(new Vec2(point[0], point[1]));
-
-      point.splice(0, 2);
-    });
-  }
-}
-
-class Line {
-
-  constructor( startX, startY, endX, endY ) {
-
-      this.startPoint = new Vec2( startX, startY );
-      this.endPoint = new Vec2( endX, endY );
-
-      this.props = {
-        visible: true
-      };
-
-      this.tags = {
-
-      };
-
-      this.type = "line";
-  }
-
-}
-
-class Box {
-
-    constructor( x, y, width, height ) {
-        
-        this.pos = {
-            x: x,
-            y: y
-        };
-        this.width = width;
-        this.height = height;
-
-        this.props = {
-
-          visible: true
-
-        };
-
-        this.tags = {
-
-        };
-
-        this.class = 'polygon';
-        this.type = 'box';
-
-        this.polygon = this.createPolygon( );
-    }
-
-    createPolygon() {
-
-        let points = [];
-
-        points.push({
-            x: this.pos.x - ( this.width / 2 ),
-            y: this.pos.y - ( this.height / 2 )
-        });
-
-        points.push({
-          x: this.pos.x + ( this.width / 2 ),
-          y: this.pos.y - ( this.height / 2 )
-        });
-
-        points.push({
-          x: this.pos.x + ( this.width / 2),
-          y: this.pos.y + ( this.height / 2)
-        });
-
-        points.push({
-          x: this.pos.x - ( this.width / 2 ),
-          y: this.pos.y + ( this.height / 2 )
-        });
-
-        return points
-    }
-
-    update() {
-
-      this.polygon = this.createPolygon( );
-
-    }
-  }
-
-export { Box, Line, Polygon, Renderer, Vec2 };
-//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoic3VuZHJvcC5tb2R1bGUuanMiLCJzb3VyY2VzIjpbXSwic291cmNlc0NvbnRlbnQiOltdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiIn0=
+export { Renderer };
